refactor: Modify EnemyShipFormation
The previous code managed enemyShips as a 2D List and operated by removing destroyed enemies from the list. To implement a bomb that destroys enemies within a 3x3 range, it was necessary to maintain the original structure of enemyShips. Therefore, instead of immediately removing destroyed enemies, the approach was changed to replace them with null. As a result, null checks have been added to the code that iterates over enemyShips.